Move useCustomerData to utils scripts to be used by multiple checkout pages
